+++ packages/gamut-styles/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [17.12.0](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.11.5...@codecademy/gamut-styles@17.12.0) (2026-03-12)
+
+### Features
+
+- Logical properties ([4c340c3](https://github.com/Codecademy/gamut/commit/4c340c3949bad30fce4f9ffabb2018763df5c4ba))
+
 ### [17.11.5](https://github.com/Codecademy/gamut/compare/@codecademy/gamut-styles@17.11.4...@codecademy/gamut-styles@17.11.5) (2026-03-04)
 
 ### Bug Fixes