+++ packages/variance/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.26.0](https://github.com/Codecademy/gamut/compare/@codecademy/variance@0.25.2...@codecademy/variance@0.26.0) (2026-03-12)
+
+### Features
+
+- Logical properties ([4c340c3](https://github.com/Codecademy/gamut/commit/4c340c3949bad30fce4f9ffabb2018763df5c4ba))
+
 ### [0.25.2](https://github.com/Codecademy/gamut/compare/@codecademy/variance@0.25.1...@codecademy/variance@0.25.2) (2026-01-28)
 
 **Note:** Version bump only for package @codecademy/variance